backend/requirements.txt
django
gunicorn
psycopg2-binary
djangorestframework
django-cors-headers
Pillow
boto3
